Transfer to JPA repositories

+++ src/main/java/com/sergosoft/goodscatalog/controller/CategoryController.java
@@ -3,15 +3,12 @@ package com.sergosoft.goodscatalog.controller;
 import com.sergosoft.goodscatalog.mapper.CategoryMapper;
 import com.sergosoft.goodscatalog.model.Category;
 import com.sergosoft.goodscatalog.service.CategoryService;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
-
 import java.util.List;
-import java.util.Optional;
 
 @Controller
 @RequestMapping("/categories")
@@ -33,7 +30,7 @@ public class CategoryController {
     }
 
     @GetMapping("/{categoryId}")
-    public String getCategoryById(@PathVariable Long categoryId, Model model) {
+    public String getCategoryById(@PathVariable Integer categoryId, Model model) {
         Category category = categoryService.getCategoryById(categoryId);
         model.addAttribute("category", categoryMapper.toDto(category));
         return "category";
